move creating all setting files into commons
